주문 내역 진행중

+++ app/src/main/java/com/nemodream/bangkkujaengi/customer/ui/viewmodel/OrderHistoryViewModel.kt
@@ -10,6 +10,6 @@ class OrderHistoryViewModel() : ViewModel() {
     var order_history_product_list = MutableLiveData<List<Purchase>>()
 
     // 구매 날짜 리스트
-    var order_history_date_list = MutableLiveData<List<Int>>()
+    var order_history_date_list = MutableLiveData<List<Long>>()
 
 }
